Eliminar transferencia

+++ FinanzasPersonales.Application/Movements/Commands/DeleteMovement/DeleteMovementCommandHandler.cs
@@ -6,10 +6,12 @@ namespace FinanzasPersonales.Application.Movements.Commands.DeleteMovement;
 public class DeleteMovementCommandHandler : IRequestHandler<DeleteMovementCommand, Unit>
 {
     private readonly IMovementRepository _movementRepository;
+    private readonly ITransferRepository _transferRepository;
 
-    public DeleteMovementCommandHandler(IMovementRepository movementRepository)
+    public DeleteMovementCommandHandler(IMovementRepository movementRepository, ITransferRepository transferRepository)
     {
         _movementRepository = movementRepository;
+        _transferRepository = transferRepository;
     }
 
     public async Task<Unit> Handle(DeleteMovementCommand request, CancellationToken cancellationToken)
@@ -21,6 +23,13 @@ public class DeleteMovementCommandHandler : IRequestHandler<DeleteMovementComman
         if (movement is null)
         {
             throw new Exception("No existe el movimiento");
+        }
+
+        // Comprobar si el movimiento es de tipo transferencia
+
+        if (_transferRepository.GetTransfersByMovementId(movement.Id.Value) is not null)
+        {
+            throw new Exception("Para eliminar este movimiento, elimine la transferencia asociada");
         }
 
         _movementRepository.Delete(movement);
